Mapear la tabla compras_productos que tiene clave compuesta

+++ platzi-market/src/main/java/com/plazi/market/persistencia/entity/Cliente.java
@@ -15,7 +15,7 @@ public class Cliente {
 
     private String apellidos;
 
-    private Double celular;
+    private Long  celular;
 
     private String direccion;
